move calendar schemas to modal dir
