+++ frontend/lib/services/google_calender/calender_service.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
-import 'package:orot/services/google_calender/schemas/event.dart';
-import 'package:orot/services/google_calender/schemas/event_attended.dart';
+import 'package:orot/modal/calendar/event.dart';
+import 'package:orot/modal/calendar/event_attended.dart';
 
 const String googleCalenderAPI = "https://www.googleapis.com/calender/v3";
